requirements.txt
matplotlib==3.2.0
Pillow==5.4.1